Fixed erronious int casting

+++ CoreSDK/Java/src/swingame/samples/IntegerSortAnimator.java
@@ -61,7 +61,7 @@ public class IntegerSortAnimator
     {
         Graphics.clearScreen();
 
-        int w = (int)(Core.screenWidth() / _data.length);
+        int w = Core.screenWidth() / _data.length;
         float scaleHeight = Core.screenHeight() / (float)_max;
 
         for(int i = 0; i < _data.length; i++)
